crud operations

+++ database/migrations/2025_09_03_182048_create_promotion_videos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('video_path')->nullable();
             $table->string('youtube_link')->nullable();
             $table->date('showing_date')->nullable();
-            $table->tinyInteger('video_order')->default(1); // 1 to 4 inside session
+            $table->integer('video_order')->default(1); // 1 to 4 inside session
             $table->tinyInteger('session_type')->default(1);  // ex: 1-"12am-12pm",2-"12pm-1pm"        
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
